+++ src/App.tsx
@@ -7,7 +7,7 @@ import {
   redirect,
 } from "react-router";
 import { Home } from "./Components/Pages/Home/Home";
-import { Form } from "./Components/Pages/Form/Form";
+import { Survey } from "./Components/Pages/Form/Survey";
 import { Results } from "./Components/Pages/Results/Results";
 import { APIDocs } from "./Components/Pages/APIDocs/APIDocs";
 import { ConfirmAddress } from "./Components/Pages/ConfirmAddress/ConfirmAddress";
@@ -108,8 +108,8 @@ const router = createBrowserRouter(
         loader={LoadAddressAndUserSession}
       />
       <Route
-        path="form"
-        element={<Form />}
+        path="survey"
+        element={<Survey />}
         loader={LoadAddressAndUserSession}
       />
       <Route path="results" element={<Results />} loader={LoadURLSession} />
