+++ src/Components/Pages/Home/Home.tsx
@@ -4,7 +4,7 @@ import { Button } from "@justfixnyc/component-library";
 
 import { GeoSearchInput } from "../../GeoSearchInput/GeoSearchInput";
 import { useSessionStorage } from "../../../hooks/useSessionStorage";
-import { FormFields } from "../Form/Form";
+import { FormFields } from "../Form/Survey";
 import { useSendGceData } from "../../../api/hooks";
 import { GCEPostData, GCEUser } from "../../../types/APIDataTypes";
 import { Header } from "../../Header/Header";
